+++ frontend/src/componentes/paginas/index.ts
@@ -17,3 +17,4 @@ export { PaginaFutbol } from './PaginaFutbol';
 export { AnalisisPartidoFutbol } from './AnalisisPartidoFutbol';
 export { BitacoraFutbol } from './BitacoraFutbol';
 export { DashboardFutbol } from './DashboardFutbol';
+export { PaginaLogin } from './PaginaLogin';